Rename ‘sources.py’ to ‘models.py’
