+++ demos/iterated-drawing/models.py
